Refactor chat route composition
