+++ src/app/(authenticated)/chat/[sessionId]/_components/chat-agent.ts
@@ -1,0 +1,3 @@
+import type { EveMessageData, UseEveAgentHelpers } from "eve/react";
+
+export type ChatAgent = UseEveAgentHelpers<EveMessageData>;
