+++ src/app/(authenticated)/chat/[sessionId]/_components/conversation/message/index.test.tsx
@@ -1,0 +1,135 @@
+import type { EveMessage } from "eve/react";
+import { renderToStaticMarkup } from "react-dom/server";
+import { describe, expect, it } from "vitest";
+import { AgentMessage } from ".";
+
+describe("agent messages", () => {
+  it("renders ordinary assistant text without a delivery tool result", () => {
+    const message = {
+      id: "assistant-message",
+      metadata: { status: "complete" },
+      parts: [
+        {
+          state: "done",
+          text: "Hello from ordinary assistant output.",
+          type: "text",
+        },
+      ],
+      role: "assistant",
+    } satisfies EveMessage;
+
+    const markup = renderToStaticMarkup(
+      <AgentMessage
+        canRespond
+        isStreaming={false}
+        message={message}
+        onInputResponses={() => undefined}
+      />
+    );
+
+    expect(markup).toContain("Hello from ordinary assistant output.");
+  });
+
+  it("renders only Linq-delivered content in the iMessage view", () => {
+    const message = {
+      id: "turn-1:assistant",
+      metadata: { status: "complete", turnId: "turn-1" },
+      parts: [
+        {
+          state: "done",
+          stepIndex: 1,
+          text: "I’ll check that now.",
+          type: "text",
+        },
+        {
+          state: "done",
+          stepIndex: 0,
+          text: "Private reasoning",
+          type: "reasoning",
+        },
+        {
+          input: { query: "example" },
+          output: { result: "internal" },
+          state: "output-available",
+          stepIndex: 0,
+          toolCallId: "call-1",
+          toolName: "web_search",
+          type: "dynamic-tool",
+        },
+        {
+          state: "done",
+          stepIndex: 1,
+          text: "Here’s what I found.",
+          type: "text",
+        },
+      ],
+      role: "assistant",
+    } satisfies EveMessage;
+
+    const markup = renderToStaticMarkup(
+      <AgentMessage
+        canRespond
+        deliveredAssistantMessages={new Map([[1, ["Here’s what I found."]]])}
+        isStreaming={false}
+        message={message}
+        onInputResponses={() => undefined}
+        userVisibleOnly
+      />
+    );
+
+    expect(markup).toContain("Here’s what I found.");
+    expect(markup).not.toContain("I’ll check that now.");
+    expect(markup).not.toContain("Private reasoning");
+    expect(markup).not.toContain("web_search");
+  });
+
+  it("shows approval controls without the hidden tool trace", () => {
+    const message = {
+      id: "turn-2:assistant",
+      metadata: { status: "streaming", turnId: "turn-2" },
+      parts: [
+        {
+          approval: { id: "approval-1" },
+          input: { amount: 50, recipient: "Hidden recipient" },
+          state: "approval-requested",
+          stepIndex: 0,
+          toolCallId: "call-2",
+          toolMetadata: {
+            eve: {
+              inputRequest: {
+                kind: "tool-approval",
+                options: [
+                  { id: "approve", label: "Approve", style: "primary" },
+                  { id: "cancel", label: "Cancel", style: "danger" },
+                ],
+                prompt: "Approve this action?",
+                requestId: "approval-1",
+              },
+              kind: "tool-call",
+              name: "send_payment",
+            },
+          },
+          toolName: "send_payment",
+          type: "dynamic-tool",
+        },
+      ],
+      role: "assistant",
+    } satisfies EveMessage;
+
+    const markup = renderToStaticMarkup(
+      <AgentMessage
+        canRespond
+        isStreaming={false}
+        message={message}
+        onInputResponses={() => undefined}
+        userVisibleOnly
+      />
+    );
+
+    expect(markup).toContain("Approve this action?");
+    expect(markup).toContain("Approve");
+    expect(markup).toContain("Cancel");
+    expect(markup).not.toContain("send_payment");
+    expect(markup).not.toContain("Hidden recipient");
+  });
+});
