+++ src/app/(authenticated)/chat/[sessionId]/_components/conversation/message/types.ts
@@ -1,0 +1,5 @@
+import type { InputResponse } from "eve/client";
+
+export type RespondToAgentInput = (
+  responses: readonly InputResponse[]
+) => void | Promise<void>;
