+++ src/app/(authenticated)/chat/[sessionId]/_lib/turn-failure.ts
