--- src/app/(authenticated)/chat/_components/agent-message.tsx
@@ -1,588 +1,0 @@
-"use client";
-
-import type {
-  EveAuthorizationPart,
-  EveDynamicToolPart,
-  EveMessage,
-  EveMessageInputRequest,
-  EveMessagePart,
-} from "eve/react";
-import {
-  CheckCircleIcon,
-  ExternalLinkIcon,
-  FileIcon,
-  ImageIcon,
-  KeyRoundIcon,
-  XCircleIcon,
-} from "lucide-react";
-import { useState } from "react";
-import {
-  Message,
-  MessageContent,
-  MessageResponse,
-} from "@/components/ai-elements/message";
-import {
-  Question,
-  QuestionActions,
-  QuestionDescription,
-  QuestionInput,
-  QuestionOption,
-  QuestionOptions,
-  QuestionPrompt,
-  type QuestionResponse,
-  QuestionSubmit,
-} from "@/components/ai-elements/question";
-import {
-  Reasoning,
-  ReasoningContent,
-  ReasoningTrigger,
-} from "@/components/ai-elements/reasoning";
-import {
-  Tool,
-  ToolContent,
-  ToolHeader,
-  ToolInput,
-  ToolOutput,
-} from "@/components/ai-elements/tool";
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
-import { cn } from "@/lib/utils";
-
-export interface AgentInputResponse {
-  readonly optionId?: string;
-  readonly requestId: string;
-  readonly text?: string;
-}
-
-type EveFilePart = Extract<EveMessagePart, { type: "file" }>;
-
-export function AgentMessage({
-  canRespond,
-  deliveredAssistantMessages,
-  isStreaming,
-  message,
-  onInputResponses,
-  timestamp,
-  userVisibleOnly = false,
-}: {
-  readonly canRespond: boolean;
-  readonly deliveredAssistantMessages?: ReadonlyMap<number, readonly string[]>;
-  readonly isStreaming: boolean;
-  readonly message: EveMessage;
-  readonly onInputResponses: (
-    responses: readonly AgentInputResponse[]
-  ) => void | Promise<void>;
-  readonly timestamp?: string;
-  readonly userVisibleOnly?: boolean;
-}) {
-  const [optimisticTimestamp] = useState(() => new Date().toISOString());
-  const displayedTimestamp =
-    timestamp ?? (message.role === "user" ? optimisticTimestamp : undefined);
-  const visibleParts = userVisibleOnly
-    ? userVisibleParts(message, deliveredAssistantMessages)
-    : message.parts;
-  const lastTextIndex = visibleParts.reduce(
-    (last, part, index) => (part.type === "text" ? index : last),
-    -1
-  );
-  const hasAssistantText =
-    message.role === "assistant" &&
-    visibleParts.some((part) => part.type === "text" && part.text.length > 0);
-
-  if (visibleParts.length === 0) return null;
-
-  return (
-    <Message
-      data-optimistic={message.metadata?.optimistic ? "true" : undefined}
-      from={message.role}
-    >
-      <MessageContent>
-        {visibleParts.map((part, index) =>
-          hasAssistantText && part.type === "reasoning" ? null : (
-            <AgentMessagePart
-              canRespond={canRespond}
-              key={partKey(part, index)}
-              onInputResponses={onInputResponses}
-              part={part}
-              showCaret={
-                isStreaming &&
-                message.role === "assistant" &&
-                index === lastTextIndex
-              }
-              userVisibleOnly={userVisibleOnly}
-            />
-          )
-        )}
-      </MessageContent>
-      {displayedTimestamp ? (
-        <time
-          className={cn(
-            "text-muted-foreground",
-            message.role === "user" ? "ml-auto pr-1" : "mr-auto"
-          )}
-          dateTime={displayedTimestamp}
-          title={formatFullTimestamp(displayedTimestamp)}
-        >
-          <span className="type-caption" suppressHydrationWarning>
-            {formatTimestamp(displayedTimestamp)}
-          </span>
-        </time>
-      ) : null}
-    </Message>
-  );
-}
-
-function userVisibleParts(
-  message: EveMessage,
-  deliveredAssistantMessages?: ReadonlyMap<number, readonly string[]>
-) {
-  if (message.role === "user")
-    return message.parts.filter(
-      (part) => part.type === "text" || part.type === "file"
-    );
-
-  const remainingDeliveries = new Map(
-    [...(deliveredAssistantMessages ?? [])].map(([stepIndex, messages]) => [
-      stepIndex,
-      [...messages],
-    ])
-  );
-
-  return message.parts.filter((part) => {
-    if (part.type === "text" && part.stepIndex !== undefined) {
-      const deliveries = remainingDeliveries.get(part.stepIndex);
-      const deliveryIndex = deliveries?.indexOf(part.text) ?? -1;
-      if (deliveryIndex < 0 || !deliveries) return false;
-      deliveries.splice(deliveryIndex, 1);
-      return true;
-    }
-
-    if (part.type === "authorization") return true;
-
-    return (
-      part.type === "dynamic-tool" &&
-      part.toolMetadata?.eve?.inputRequest !== undefined
-    );
-  });
-}
-
-const timestampFormatter = new Intl.DateTimeFormat(undefined, {
-  hour: "numeric",
-  minute: "2-digit",
-});
-
-const fullTimestampFormatter = new Intl.DateTimeFormat(undefined, {
-  dateStyle: "medium",
-  timeStyle: "short",
-});
-
-function formatTimestamp(timestamp: string) {
-  return timestampFormatter.format(new Date(timestamp));
-}
-
-function formatFullTimestamp(timestamp: string) {
-  return fullTimestampFormatter.format(new Date(timestamp));
-}
-
-function AgentMessagePart({
-  canRespond,
-  onInputResponses,
-  part,
-  showCaret,
-  userVisibleOnly,
-}: {
-  readonly canRespond: boolean;
-  readonly onInputResponses: (
-    responses: readonly AgentInputResponse[]
-  ) => void | Promise<void>;
-  readonly part: EveMessagePart;
-  readonly showCaret: boolean;
-  readonly userVisibleOnly: boolean;
-}) {
-  switch (part.type) {
-    case "step-start":
-      return null;
-    case "text":
-      return (
-        <MessageResponse caret="block" isAnimating={showCaret}>
-          {part.text}
-        </MessageResponse>
-      );
-    case "reasoning":
-      return (
-        <Reasoning defaultOpen isStreaming={part.state === "streaming"}>
-          <ReasoningTrigger />
-          <ReasoningContent>{part.text}</ReasoningContent>
-        </Reasoning>
-      );
-    case "file":
-      return <AttachmentPart part={part} />;
-    case "authorization":
-      return <AuthorizationPrompt part={part} />;
-    case "dynamic-tool": {
-      const inputRequest = part.toolMetadata?.eve?.inputRequest;
-      if (inputRequest?.kind === "question") {
-        return (
-          <QuestionRequest
-            canRespond={canRespond}
-            inputRequest={inputRequest}
-            inputResponse={part.toolMetadata?.eve?.inputResponse}
-            onInputResponses={onInputResponses}
-          />
-        );
-      }
-
-      if (userVisibleOnly && inputRequest) {
-        return (
-          <InputRequestActions
-            canRespond={canRespond}
-            part={part}
-            onInputResponses={onInputResponses}
-          />
-        );
-      }
-
-      const tool = (
-        <Tool
-          defaultOpen={
-            part.state === "approval-requested" ||
-            part.state === "approval-responded"
-          }
-        >
-          <ToolHeader status={part.state} title={part.toolName} />
-          <ToolContent>
-            <ToolInput input={part.input} />
-            <InputRequestActions
-              canRespond={canRespond}
-              part={part}
-              onInputResponses={onInputResponses}
-            />
-            <ToolOutput errorText={part.errorText} output={part.output} />
-          </ToolContent>
-        </Tool>
-      );
-      return tool;
-    }
-  }
-  throw new Error("Unsupported agent message part.");
-}
-
-function QuestionRequest({
-  canRespond,
-  inputRequest,
-  inputResponse,
-  onInputResponses,
-}: {
-  readonly canRespond: boolean;
-  readonly inputRequest: EveMessageInputRequest;
-  readonly inputResponse?: AgentInputResponse;
-  readonly onInputResponses: (
-    responses: readonly AgentInputResponse[]
-  ) => void | Promise<void>;
-}) {
-  const selectedOption = inputRequest.options?.find(
-    (option) => option.id === inputResponse?.optionId
-  );
-  const hasOptions = (inputRequest.options?.length ?? 0) > 0;
-  const acceptsFreeform = inputRequest.allowFreeform === true || !hasOptions;
-
-  const submitResponse = ({ selectedValues, text }: QuestionResponse) =>
-    onInputResponses([
-      {
-        optionId: selectedValues[0],
-        requestId: inputRequest.requestId,
-        text,
-      },
-    ]);
-
-  return (
-    <Question
-      defaultValue={{
-        selectedValues: inputResponse?.optionId ? [inputResponse.optionId] : [],
-        text: inputResponse?.text ?? "",
-      }}
-      disabled={!canRespond || inputResponse !== undefined}
-      onSubmit={submitResponse}
-    >
-      <QuestionPrompt>{inputRequest.prompt}</QuestionPrompt>
-      {hasOptions ? (
-        <QuestionOptions
-          className="flex-col items-stretch"
-          aria-label={inputRequest.prompt}
-        >
-          {inputRequest.options?.map((option) => (
-            <QuestionOption
-              className="justify-start text-left"
-              key={option.id}
-              value={option.id}
-            >
-              <span>
-                <span className="block">{option.label}</span>
-                {option.description ? (
-                  <span className="block type-caption opacity-70">
-                    {option.description}
-                  </span>
-                ) : null}
-              </span>
-            </QuestionOption>
-          ))}
-        </QuestionOptions>
-      ) : null}
-      {acceptsFreeform ? (
-        <QuestionInput aria-label="Answer" placeholder="Type your answer…" />
-      ) : null}
-      {inputResponse ? (
-        <QuestionDescription>
-          Responded:{" "}
-          {selectedOption?.label ??
-            inputResponse.text ??
-            inputResponse.optionId}
-        </QuestionDescription>
-      ) : (
-        <QuestionActions>
-          <QuestionSubmit>Answer</QuestionSubmit>
-        </QuestionActions>
-      )}
-    </Question>
-  );
-}
-
-function AttachmentPart({ part }: { readonly part: EveFilePart }) {
-  const label = part.filename ?? "Attachment";
-  const detail = [part.mediaType, formatBytes(part.size)]
-    .filter(Boolean)
-    .join(" · ");
-  const isImage = part.mediaType.startsWith("image/") && part.url !== undefined;
-  const Icon = isImage ? ImageIcon : FileIcon;
-  const content = (
-    <>
-      {isImage ? (
-        // Browser artifacts use runtime URLs that cannot be declared in Next Image configuration.
-        // oxlint-disable-next-line nextjs/no-img-element -- runtime browser artifact URL
-        <img
-          alt={label}
-          className="size-12 shrink-0 rounded-sm object-cover"
-          src={part.url}
-        />
-      ) : (
-        <span className="flex size-10 shrink-0 items-center justify-center rounded-sm bg-muted text-muted-foreground">
-          <Icon className="size-4" />
-        </span>
-      )}
-      <span className="min-w-0 flex-1">
-        <span className="block truncate type-label">{label}</span>
-        {detail ? (
-          <span className="block truncate text-muted-foreground">{detail}</span>
-        ) : null}
-      </span>
-      {part.url ? (
-        <ExternalLinkIcon className="size-4 shrink-0 text-muted-foreground" />
-      ) : null}
-    </>
-  );
-
-  return part.url ? (
-    <Button
-      className="max-w-sm"
-      nativeButton={false}
-      render={
-        <a
-          aria-label={`Open ${label}`}
-          href={part.url}
-          rel="noreferrer"
-          target="_blank"
-        />
-      }
-      variant="surface"
-    >
-      {content}
-    </Button>
-  ) : (
-    <Card className="max-w-sm" size="sm">
-      <CardContent className="flex items-center gap-3">{content}</CardContent>
-    </Card>
-  );
-}
-
-function AuthorizationPrompt({
-  part,
-}: {
-  readonly part: EveAuthorizationPart;
-}) {
-  const isAuthorized =
-    part.state === "completed" && part.outcome === "authorized";
-  const isCompleted = part.state === "completed";
-  const Icon = isAuthorized
-    ? CheckCircleIcon
-    : isCompleted
-      ? XCircleIcon
-      : KeyRoundIcon;
-  const instructions = part.authorization?.instructions;
-  const shouldShowInstructions =
-    instructions !== undefined && instructions !== part.description;
-  const alertVariant = isAuthorized
-    ? "success"
-    : isCompleted
-      ? "destructive"
-      : "information";
-
-  return (
-    <Alert variant={alertVariant}>
-      <Icon />
-      <AlertTitle>{authorizationTitle(part)}</AlertTitle>
-      <AlertDescription>
-        <p>{authorizationDescription(part)}</p>
-        {shouldShowInstructions ? <p>{instructions}</p> : null}
-        {part.state === "required" && part.authorization?.userCode ? (
-          <div className="flex flex-wrap items-center gap-2">
-            <span>Code</span>
-            <Badge variant="outline">
-              <code className="type-compact-code">
-                {part.authorization.userCode}
-              </code>
-            </Badge>
-          </div>
-        ) : null}
-        {part.state === "required" && part.authorization?.url ? (
-          <Button
-            render={
-              <a
-                aria-label={`Sign in with ${part.displayName}`}
-                href={part.authorization.url}
-                rel="noreferrer"
-                target="_blank"
-              />
-            }
-            size="sm"
-          >
-            <ExternalLinkIcon />
-            Sign in with {part.displayName}
-          </Button>
-        ) : null}
-      </AlertDescription>
-    </Alert>
-  );
-}
-
-function authorizationTitle(part: EveAuthorizationPart): string {
-  if (part.state === "required") {
-    return `Connect ${part.displayName}`;
-  }
-  if (part.outcome === "authorized") {
-    return `${part.displayName} connected`;
-  }
-  return `${part.displayName} authorization ${formatAuthorizationOutcome(part.outcome)}`;
-}
-
-function authorizationDescription(part: EveAuthorizationPart): string {
-  if (part.state === "required") {
-    return part.description;
-  }
-  if (part.outcome === "authorized") {
-    return `${part.displayName} connected.`;
-  }
-  const tail = part.reason !== undefined ? ` (${part.reason})` : "";
-  return `${part.displayName} authorization ${formatAuthorizationOutcome(part.outcome)}${tail}.`;
-}
-
-function formatAuthorizationOutcome(
-  outcome: NonNullable<EveAuthorizationPart["outcome"]>
-): string {
-  switch (outcome) {
-    case "authorized":
-      return "authorized";
-    case "declined":
-      return "declined";
-    case "failed":
-      return "failed";
-    case "timed-out":
-      return "timed out";
-  }
-  throw new Error("Unsupported authorization outcome.");
-}
-
-function formatBytes(size: number | undefined): string | undefined {
-  if (size === undefined) {
-    return undefined;
-  }
-  if (size < 1024) {
-    return `${String(size)} B`;
-  }
-  if (size < 1024 * 1024) {
-    return `${(size / 1024).toFixed(1)} KB`;
-  }
-  return `${(size / (1024 * 1024)).toFixed(1)} MB`;
-}
-
-function InputRequestActions({
-  canRespond,
-  onInputResponses,
-  part,
-}: {
-  readonly canRespond: boolean;
-  readonly onInputResponses: (
-    responses: readonly AgentInputResponse[]
-  ) => void | Promise<void>;
-  readonly part: EveDynamicToolPart;
-}) {
-  const inputRequest = part.toolMetadata?.eve?.inputRequest;
-  if (!inputRequest) {
-    return null;
-  }
-
-  const inputResponse = part.toolMetadata.eve.inputResponse;
-  const selectedOption = inputRequest.options?.find(
-    (option) => option.id === inputResponse?.optionId
-  );
-
-  return (
-    <Alert variant="warning">
-      <AlertTitle>{inputRequest.prompt}</AlertTitle>
-      <AlertDescription>
-        {inputResponse ? (
-          <p>
-            Responded:{" "}
-            {selectedOption?.label ??
-              inputResponse.text ??
-              inputResponse.optionId}
-          </p>
-        ) : (
-          <div className="flex flex-wrap gap-2">
-            {inputRequest.options?.map((option) => (
-              <Button
-                disabled={!canRespond}
-                key={option.id}
-                onClick={() => {
-                  void onInputResponses([
-                    {
-                      optionId: option.id,
-                      requestId: inputRequest.requestId,
-                    },
-                  ]);
-                }}
-                size="sm"
-                type="button"
-                variant={option.style === "danger" ? "destructive" : "default"}
-              >
-                {option.label}
-              </Button>
-            ))}
-          </div>
-        )}
-      </AlertDescription>
-    </Alert>
-  );
-}
-
-function partKey(part: EveMessagePart, index: number): string {
-  switch (part.type) {
-    case "authorization":
-      return `authorization:${part.turnId}:${String(part.stepIndex)}:${part.name}`;
-    case "dynamic-tool":
-      return part.toolCallId;
-    default:
-      return `${part.type}:${String(index)}`;
-  }
-}
