+++ src/app/(authenticated)/chat/_lib/message-input.ts
@@ -1,0 +1,25 @@
+import type { UserContent } from "ai";
+import type { PromptInputMessage } from "@/components/ai-elements/prompt-input";
+
+export function messageContent(message: PromptInputMessage) {
+  const text = message.text.trim();
+  if (message.files.length === 0) return text;
+
+  const parts: UserContent = [];
+  if (text.length > 0) parts.push({ text, type: "text" });
+  for (const file of message.files) {
+    parts.push({
+      data: file.url,
+      filename: file.filename,
+      mediaType: file.mediaType,
+      type: "file",
+    });
+  }
+  return parts;
+}
+
+export function chatTitle(message: PromptInputMessage) {
+  const text = message.text.trim();
+  if (text) return text.slice(0, 240);
+  return message.files[0]?.filename?.slice(0, 240) ?? "New chat";
+}
